+++ repositories/automation-stack/images/jenkins-master/resources/user-content/theme.js
@@ -1,0 +1,3 @@
+document.addEventListener("DOMContentLoaded", function () {
+    document.title = document.title.replace("[Jenkins]", "[My Jenkins]");
+});