+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Regroup Compose files (`99-development.yml` -> `01-development.yml`).
+- Merge `php-web-server` and `web-client` repositories into `web-application-stack` monorepo.
 
 ## [0.7.7] - 2018-10-05