+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Install Configuration as Code Jenkins plugin.
+
 ## [0.8.0] - 2018-10-07
 
 ### Added