+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Install Configuration as Code Jenkins plugin.
+- Customize Jenkins theme as code.
 
 ## [0.8.0] - 2018-10-07