+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.8.1] - 2018-10-17
+
 ### Added
 
 - Install Configuration as Code Jenkins plugin.